refactor(primitives): migrate Point to tag-based coordinate system
- Replace Dimension template parameter with Tag-based layout system.
- Add support for OGC types (XY, XYZ, XYM, XYZM).
- Separate spatial_dims (for geometry logic) from storage_size (for memory).
- Update Doxygen documentation to reflect the new template structure.

BREAKING CHANGE: Point template now expects a Tag instead of a Dimension enum.

+++ src/tidop/math/algebra/decomp/svd.h
@@ -419,7 +419,7 @@ void SingularValueDecomposition<Matrix_t<T, Rows, Cols>>::decompose()
         if (with_v) {
 
             int l = -1;
-            for (int i = mCols - 1; i >= 0; --i) {
+            for (int i = static_cast<int>(mCols) - 1; i >= 0; --i) {
 
                 if (l != -1) {
